Fix(post): 메인페이지 오류 수정
