+++ src/hooks/usePostsQuery.tsx
@@ -3,6 +3,7 @@
 import { getPostList } from "@/lib/postApi"
 import { usePostsStore } from "@/stores/usePostsStore";
 import { useQuery } from "@tanstack/react-query"
+import { useEffect } from "react";
 
 interface PostType {
   id: number;
@@ -18,7 +19,12 @@ export const usePostsQuery = () => {
     queryFn: getPostList,
     // staleTime: 1000 * 60 * 5, // 쿼리 데이터가 최신 데이터로 간주되는 시간
   })
-  if (data) {
-    setPosts(data);
-  }
+  // data가 있을 때만 setPosts 호출
+  useEffect(() => {
+    if (data) {
+      setPosts(data);
+    }
+  }, [data, setPosts]);
+
+  return data;
 }
